App.js -> index.js | index.js -> entry.js
