+++ packages/create-redwood-app/template/web/src/entry.js
@@ -1,0 +1,13 @@
+import ReactDOM from 'react-dom'
+
+import App from '.'
+
+import './index.css'
+
+const rootElement = document.getElementById('redwood-app')
+
+if (rootElement.hasChildNodes()) {
+  ReactDOM.hydrate(<App />, rootElement)
+} else {
+  ReactDOM.render(<App />, rootElement)
+}
