+++ packages/create-redwood-app/template/web/src/index.js
@@ -1,13 +1,15 @@
-import ReactDOM from 'react-dom'
+import { FatalErrorBoundary } from '@redwoodjs/web'
+import { RedwoodApolloProvider } from '@redwoodjs/web/apollo'
 
-import App from './App'
+import FatalErrorPage from 'src/pages/FatalErrorPage'
+import Routes from 'src/Routes'
 
-import './index.css'
+const App = () => (
+  <FatalErrorBoundary page={FatalErrorPage}>
+    <RedwoodApolloProvider>
+      <Routes />
+    </RedwoodApolloProvider>
+  </FatalErrorBoundary>
+)
 
-const rootElement = document.getElementById('redwood-app')
-
-if (rootElement.hasChildNodes()) {
-  ReactDOM.hydrate(<App />, rootElement)
-} else {
-  ReactDOM.render(<App />, rootElement)
-}
+export default App
